requirements.txt
datasets
peft
trl
accelerate
bitsandbytes
sentencepiece
tensorboard
google-cloud-storage
google-cloud-aiplatform
gcsfs
evaluate
rouge_score
vllm